requirements.txt
altgraph==0.17.4
anyio==4.0.0
astor==0.8.1
attrdict==2.0.1
Babel==2.13.1
bce-python-sdk==0.8.92
beautifulsoup4==4.12.2
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
colorthief==0.2.1
contourpy==1.1.1
cssselect==1.2.0
cssutils==2.9.0
cycler==0.12.1
Cython==3.0.4
darkdetect==0.8.0
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.43.1
future==0.18.3
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.6
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.7.3
networkx==3.1
numpy==1.24.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle-gpu==2.5.1
pandas==2.0.3
pdf2docx==0.5.6
pefile==2023.2.7
Pillow==10.1.0
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.6
pyclipper==1.3.0.post5
pycryptodome==3.19.0
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyMuPDF==1.20.2
pyparsing==3.1.1
PyQt-Fluent-Widgets==1.3.3
PyQt5==5.15.9
PyQt5-Frameless-Window==0.3.3
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
rapidfuzz==3.4.0
rarfile==4.1
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
termcolor==2.3.0
tifffile==2023.7.10
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
visualdl==2.5.3
Werkzeug==3.0.1
zipp==3.17.0
